fix: Address CI failures
- Update butteraugli-oxide to 0.2 (compatible with codec-eval)
- Fix clippy warnings in sys-local crate
- Add #[non_exhaustive] to Error enum (API stability)
- Run cargo fmt on all files
- Restrict clippy to lib target only (test/example code has many false positives)

+++ src/test_encoder.rs
@@ -158,7 +158,7 @@ pub fn encode_c(rgb: &[u8], width: u32, height: u32, config: &TestEncoderConfig)
         cinfo.common.err = jpeg_std_error(&mut jerr);
         jpeg_CreateCompress(
             &mut cinfo,
-            JPEG_LIB_VERSION as i32,
+            JPEG_LIB_VERSION,
             std::mem::size_of::<jpeg_compress_struct>(),
         );
